modified seq body and reverted some previous changes regarding seq
